Animal entity and layers added

+++ src/main/java/dev/patika/ecommerce/business/abstracts/IAnimalService.java
@@ -8,9 +8,9 @@ import java.util.List;
 
 public interface IAnimalService {
     Animal save(Animal animal);
-    Animal get(int id);
+    Animal get(long id);
     Animal update(Animal animal);
-    boolean delete(int id);
+    boolean delete(long id);
     Page<Animal> cursor(int page, int pageSize);
     List<Animal> findByName(String name);
